requirements.txt
Django==4.2.5
django-bootstrap-v5==1.0.11
django-heroku
gunicorn
django-allauth==0.57.0
whitenoise==6.5.0
django-autoslug==1.9.9
psycopg2
python-dotenv==1.0.0
ics
pytz